Fixed error handling around config file
